Add some new TACO devices: IOCounter and IOTimer
The access to the event/monitor counters and timer is now given via separate
TACO devices. The configuration is done give in an example resource file:
qmesydaq.res

The device names are defined in the TACO section of the common configuration
file of QMesyDAQ in QMesyDAQ.ini of the Qt system.
entries are:
detector
events
timer
counter0..counter3
the TACO device names are given as value of those keys.

git-svn-id: https://trac.frm2.tum.de/qmesydaq/trunk@617 71953026-4e7b-de11-b703-00e01807f32f