Add the view to the position corrected histograms.
The include of the position corrected histograms forces a lot of
changes because the histogram was oriented in that way, that the
x axis was connected to a tube, but the display has to be rotated
90 degrees counterclockwise.

To simplify the code the MappedHistogram class is now derived from
the Histogram.

Introduce a histogram object. The histogram object stores the informations
about the data and it allows an simpler access to the data in different
display configurations.

Create a new correction map 'UserMapCorrection' to read the correction
map via the object from a file. It allows to access much easier to the
correction interface. No blowing up of the interface.

git-svn-id: https://trac.frm2.tum.de/qmesydaq/trunk@649 71953026-4e7b-de11-b703-00e01807f32f