Reorganize access to version, capabilities.
The access to the version and capabilities information of the MCPD and MPSD
has changed. During a 'scanPeriph' call of a MCPD the version and capabilities
will be read from all modules connected to the MCPD and stored in this classes
the readout of this values will be done by special methods. This avoids a lot
of hardware access and allows a more consistant access to the information.

If the hardware has changed due to adding or removing a module all the information
must be read via 'scanPeriph' method.

git-svn-id: https://trac.frm2.tum.de/qmesydaq/trunk@663 71953026-4e7b-de11-b703-00e01807f32f