Add the loading of a lookup table correction file.
Loading and using the lookup table for loading histograms, replay
list mode files, and (hopefully) also for live data aquisition.

This file format is given by the detector group of the FRM-II for
the SANS-1 machine. This file format is not perfect, but it could
be a good starting point.

Missing in the file format are the dimensions of source and target.

git-svn-id: https://trac.frm2.tum.de/qmesydaq/trunk@717 71953026-4e7b-de11-b703-00e01807f32f