Change the text at the plot buttons.
The text at the plot type selection radio buttons and the order of
the buttons has changed. The former position and amplitude buttons
are marked with 'raw' to indicate, that this are the data coming
from the hardware without any transformation.

The normal (standard) view of data should be the position (rearranged
and sensitivy corrected) data.

git-svn-id: https://trac.frm2.tum.de/qmesydaq/trunk@719 71953026-4e7b-de11-b703-00e01807f32f