FIX: timer problems
The timer sometimes stop with an huge time value.

This is caused on remaining data packages from previous measurement.
Either the data will be send by the hardware after start or there
are some in the network stack.

Current solution is to remove all packages with header times older
than start time. The writing of the listmode file has also to be changed.

git-svn-id: https://trac.frm2.tum.de/qmesydaq/trunk@730 71953026-4e7b-de11-b703-00e01807f32f