Change the start and stop of the measurement.
Due to a misunderstanding (or malfunction of hardware) the measuring
start point was not right set.

If you have a system with more than 1 MCPD the master/slave operation
did not work as expected. The slave sends data when it gets the start
command. These data have to be ignored in the hardware. There seems
no sync between master and slave in this particular case.

git-svn-id: https://trac.frm2.tum.de/qmesydaq/trunk@738 71953026-4e7b-de11-b703-00e01807f32f