External synchronisation of MCPD-8.
The needed software modifications in qmesydaq (mainly for a master
MCPD-8 has to be set another checkbox, to enable the external sync

In the protocol the UDP command 6 (setTiming) has changed:
in word 10 now are the values Werte 0, 1 und 3 (=master+external sync)
allowed.

The monitor/chopper input 3 converts to the external sync input.

Only the master is allowed to be external synced.
If the master is externally sync the monitor/chopper input 3 will be
disabled as trigger source.
Display the external/internal sync configuration in the status bar.

For the slave modules nothings changes.

Change-Id: I6dbae18412af99a0c2e204ee5b6129af2db9971b
Signed-off-by: Jens Krüger <jens.krueger@frm2.tum.de>
Reviewed-on: https://trac.frm2.tum.de/review/1863
Reviewed-by: Lutz Rossa <rossa@helmholtz-berlin.de>
Reviewed-by: Jens Krueger Jens Krüger <jens.krueger@frm2.tum.de>
Tested-by: Jens Krueger Jens Krüger <jens.krueger@frm2.tum.de>