fix race condition in logging handlers
The logging handler had a race condition, if a context switch of one thread
inside the logging handler has occured and a second thread did some logging
and the millisecond time was different to thread one. Now every thread
generates a own time string for logging and the is a mutex for writing to
the log file.

Change-Id: I78078acf31b0fcff76e7fb10dc7b653998e35d77
Signed-off-by: Lutz Rossa <rossa@helmholtz-berlin.de>
Reviewed-on: https://trac.frm2.tum.de/review/3830
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>