FIXES: The switching between position and amplitude
The switches between amplitude/energy and position did not work as
expected. In the MCPD setup it was not implemented and for th modules
itself, it only worked by changing the toggle buttons. This was not
very intuitive.

Now a button "Set Mode" is added to set the value explictly.
Additionally a checkbox allows the setting of the mode for all MPSD at
ones.

Change-Id: I4a0d68fe94aada8cd5b3457f90af1a6192bfa65a
fixes: #1157,#1149
Reviewed-on: https://trac.frm2.tum.de/review/4082
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>