Remote configuration file loading
The configuration file could be loaded remotely. The configuration
file name will be given and it will search in the default path for
the configuration files. If found and it is a valid file, it will
be loaded.

The TACO resource is called 'configfile'.

If it will be queried the file without a path will given.

Fixes: #204
Change-Id: I82d54852bd7171b8dd14632d470bdaac9caf3405
Reviewed-on: https://forge.frm2.tum.de/review/4181
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>