Sane handling of different setups with different modules
If the hardware setup has connection to different modules and the
configuration files handle only some of the modules the other modules
should not influence the measurement setup.

At PUMA instrument we have a MPSD-8 and a MDST-16 module connected to
the MCPD-8. It should be possible to switch between two different setups
without changing the hardware setup (remove and plug cable, ... )

Change-Id: Ifbfa3291ec6e7df3209d255bd6312c0793aa8834
Fixes: #1571
Reviewed-on: https://forge.frm2.tum.de/review/6054
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>