LIB: Shift the pos/amp of old MPSD neutron events earlier
The position/amplitude of the neutron events of the old MPSD8 modules
(SPODI type) are left aligned in the 10 bits and therefore they must
be shifted bit 2 bits to the right. This was done to late, so that
some of the events were sorted out if their position amplitude was
greater then 240. This became relevant for the 'right' pulser events.

Change-Id: I7de2cd81653e4100562b04746cb61b3000640024
Reviewed-on: https://forge.frm2.tum.de/review/6450
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>