allow automatic file wrapping to next file
To prevent too big files, the interface is able to set a file size, which
triggers a switch to the next output file. On success, QMesyDAQ will write an
end separator and close the older output file. The newer output file will get
a header and a starting separator. On errors, QMesyDAQ will switch back to
one big file.

You can use a printf-like syntax for file names:
  %r  run number          (default 8 digits)
  %p  file part number (starting with 1, default 3 digits)
  %H  hour         (0..23, default 2 digits)
  %M  minute              (default 2 digits)
  %S  second              (default 2 digits)
  %d  day of month (1..31, default 2 digits)
  %m  month        (1..12, default 2 digits)
  %y  year                (default 4 digits)
  %%  the percent sign
other formats are ignored and part of the file name and you are allowed to
insert a field width (padded with '0'),	e.g. "%2p" will be replaced by
"01", "02", "03" ...

Change-Id: I9c75b30ddaf1cadaca833455dd4c76a18bdd5def
Signed-off-by: Lutz Rossa <rossa@helmholtz-berlin.de>
Reviewed-on: https://forge.frm2.tum.de/review/7971
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>