Introduce a new histogram file format: Simple Histogram File Format
For some reason it is nice to have a different histogram file format.
This could be selected via the general settings dialog.

The new format called 'Simple Histogram File Format' is the same
format which was used a the SPHERES instrument. It only contains the
pure uncorrected position histogram data without any additional
entries and no column or line numbers.

Change-Id: I7659c4fed1e5461c43293a25e595aa33e7cf1da9
Reviewed-on: https://forge.frm2.tum.de/review/8347
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>