FIX: Settings and display of common gain
The detection of common gain is now done by the contents of the gain
cells itself. The algorithm uses only the active cells of a module to
detect the common gain. If all have the same value then this value
will given back otherwise the value of the 'common' channel cell.
Fixes: #2109

Change-Id: I07604dd1365501873f038cc2459fe5ff1f093ea4
Reviewed-on: https://forge.frm2.tum.de/review/8449
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jens Krueger <jens.krueger@frm2.tum.de>