Add a comment into the config file during the save dialog
The file saving dialog for the configuration files was extended by a
text field to add a comment, which will be stored inside the
configuration file in QMesyDAQ/comment. This could be also displayed
in the QMesyDAQ frontend in the files section.

Additionally this could be extended to store some comments in listmode
or history files if they are stored manually.

Change-Id: If0f40600f4ee002b8b0f2c08fb590de53ad17d42
Reviewed-on: https://forge.frm2.tum.de/review/8709
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>