LIB: detection of mapping size from LUT calibration file
The detection of the source and destination size for the calibration
via the LUT (look up table) calibration file format is done from the
file itself. In the 3rd header line the number of tubes is given
wereas the number of bins in a tube is given in the 1st column. The
destination width is also given by the tubes and the destination
height is detected by the maximum number of the 'pix' entries.

Change-Id: I6e782ffc621baf8d4b618dfe88f8c098c86bad50
Reviewed-on: https://forge.frm2.tum.de/review/8853
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>