Add two new monitors to the main widget
The two TTL inputs of the modules could now be used as monitors too.
The counter for monitors are separated now from the event counter and
the timer. The do not belong any longer to the counter list. So we
could manage the monitors from the different modules a little bit better
in the future.

Change-Id: I502785bdf8f082d8899e8032161e98477b681868
Reviewed-on: https://forge.frm2.tum.de/review/11006
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>